requirements.txt
Django
djangorestframework
matplotlib
